EDIT: Forgot method on Fake Repo and Typo in MyPDO Repo and Change URLs
Forgot to add the methods in the Fake Repo so it was throwing errors when it was launching. These have now been resolved.

Mistyped the column name in the methods in the MyPDO Repo. These have now been changed and the SQL successful updates now.

Changed the URL's for the enable and disbale on the Admin Index page to link to the product ID. This is due to the routing expecting a id after the action.

[+] Add methods to Fake Repo
[+] Fix typo in MyPdo Repo
[+] Adjust url in Admin index for the router