EDIT: Using Product Model on Admin Index results
Running of the products through the Product Model on the Admin Index Page to allow me to build up an object with the Categories that I can use at a later date.

This will be used on all pages and I can then use it to pretick the categories selected on the edit pages.

[+] Products using the Model and adding the Categories